improved UnitTests for vendors MouldKing and JIESTAR by using AssemblyModules-DI
